# api/src/main/api/SettingsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.ben_kostka.benchesster.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-12-31T20:03:31.088752+01:00[Europe/Berlin]", comments = "Generator version: 7.10.0")
@Validated
@Tag(name = "settings", description = "the settings API")
public interface SettingsApi {
    /**
     * GET /settings : See options panel
     * Returns all possible settings options
     *
     * @return successful operation (status code 200)
     *         or Invalid ID supplied (status code 400)
     *         or Account not found (status code 404)
     */
    @Operation(
        operationId = "getSettingsoptions",
        summary = "See options panel",
        description = "Returns all possible settings options",
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "text/*", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid ID supplied"),
            @ApiResponse(responseCode = "404", description = "Account not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/settings",
        produces = { "text/*" }
    )
    
    public ResponseEntity<String> getSettingsoptions();

}